+++ lib/currency_widget.dart
@@ -1,6 +1,7 @@
 library currency_widget;
-export  'package:currency_widget/src/models/Currency.dart';
+export 'package:currency_widget/src/models/currency.dart';
 export 'package:currency_widget/src/view/currency_picker.dart';
 export 'package:currency_widget/src/Controller/currency_controller.dart';
 export 'package:currency_widget/src/view/currency_textview.dart';
 export 'package:currency_widget/src/view/currency_textfield.dart';
+export 'package:currency_widget/src/view/currency_card_view_report.dart';